move files to directories where they belong to
