+++ Software/src/lib/eModbus-eModbus/scripts/README.md
@@ -1,0 +1,1 @@
+The scripts in this folder originate from [eModbus discussion 147](https://github.com/eModbus/eModbus/discussions/147).
